chg: updated Entities with Lombok and small structural refactor
