+++ src/main/java/sk/leziemevpezinku/spring/api/validation/Enumeration.java
@@ -1,4 +1,4 @@
-package sk.leziemevpezinku.spring.model.annotation.validation;
+package sk.leziemevpezinku.spring.api.validation;
 
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
