+++ src/main/java/sk/leziemevpezinku/spring/api/validation/FreeText.java
@@ -1,4 +1,4 @@
-package sk.leziemevpezinku.spring.model.annotation.validation;
+package sk.leziemevpezinku.spring.api.validation;
 
 import jakarta.validation.constraints.Pattern;
