+++ src/main/java/sk/leziemevpezinku/spring/model/Attendance.java
@@ -1,29 +1,34 @@
 package sk.leziemevpezinku.spring.model;
 
 import jakarta.persistence.*;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
+import lombok.*;
 import sk.leziemevpezinku.spring.model.base.AuditedEntityBase;
 import sk.leziemevpezinku.spring.model.enums.AttendanceStatus;
 
 @Data
-@EqualsAndHashCode(callSuper = true)
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
 @Entity
 @Table(name = "reg_attendance")
 public class Attendance extends AuditedEntityBase {
 
+    @ToString.Include
     @EmbeddedId
     private AttendanceId id;
 
+    @EqualsAndHashCode.Exclude
     @MapsId("lessonId")
     @ManyToOne(fetch = FetchType.LAZY)
     private Lesson lesson;
 
+    @EqualsAndHashCode.Exclude
     @MapsId("personId")
     @ManyToOne(fetch = FetchType.LAZY)
     private Person person;
 
     /** Enumerated :: {@link AttendanceStatus} */
+    @ToString.Include
     @Column(name = "status", length = 10, nullable = false)
     private String status;
 }
