+++ src/main/java/sk/leziemevpezinku/spring/model/Lesson.java
@@ -1,8 +1,7 @@
 package sk.leziemevpezinku.spring.model;
 
 import jakarta.persistence.*;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
+import lombok.*;
 import org.hibernate.annotations.JdbcTypeCode;
 import org.hibernate.type.SqlTypes;
 import sk.leziemevpezinku.spring.model.base.AuditedEntityBase;
@@ -12,31 +11,40 @@ import java.util.ArrayList;
 import java.util.List;
 
 @Data
-@EqualsAndHashCode(callSuper = true)
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
 @Entity
 @Table(name = "reg_lesson")
 public class Lesson extends AuditedEntityBase {
 
     @Id
+    @ToString.Include
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seq_lesson_id")
     @SequenceGenerator(name = "seq_lesson_id", sequenceName = "seq_lesson_id", initialValue = 1, allocationSize = 1)
     private Long id;
 
+    @ToString.Include
     @Column(name = "start_at", nullable = false)
     private LocalDateTime startAt;
 
+    @ToString.Include
     @Column(name = "duration_minutes", nullable = false)
     private Integer durationMinutes;
 
+    @Builder.Default
+    @EqualsAndHashCode.Exclude
     @JdbcTypeCode(value = SqlTypes.JSON)
     @Column(name = "expected_trainers")
     private List<String> expectedTrainers = new ArrayList<>();
 
+    @EqualsAndHashCode.Exclude
     @ManyToOne(optional = false)
     @JoinColumn(name = "place_id", nullable = false)
     private Place place;
 
+    @EqualsAndHashCode.Exclude
     @ManyToOne(optional = false)
     @JoinColumn(name = "event_term_id", nullable = false)
     private EventTerm eventTerm;
