+++ src/main/java/sk/leziemevpezinku/spring/model/Registration.java
@@ -10,22 +10,26 @@ import java.util.List;
 
 @Data
 @Builder
-@EqualsAndHashCode(callSuper = true)
+@NoArgsConstructor
+@AllArgsConstructor
 @Entity
 @Table(name = "reg_registration")
 public class Registration extends CreatedAtEntityBase {
 
     @Id
+    @ToString.Include
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seq_registration_id")
     @SequenceGenerator(name = "seq_registration_id", sequenceName = "seq_registration_id", initialValue = 1, allocationSize = 1)
     private Long id;
 
+    @ToString.Include
     @Column(name = "uuid", length = 40, nullable = false, updatable = false)
     private String uuid;
 
     /** Enumerated :: {@link RegistrationStatus} */
     @Builder.Default
+    @ToString.Include
     @Column(name = "status", length = 10, nullable = false)
     private String status = RegistrationStatus.CONCEPT.name();
 
@@ -50,15 +54,18 @@ public class Registration extends CreatedAtEntityBase {
     @Column(name = "consent_photo")
     private Boolean consentPhoto;
 
+    @EqualsAndHashCode.Exclude
     @ManyToOne(optional = false)
     @JoinColumn(name = "event_term_id", nullable = false)
     private EventTerm eventTerm;
 
+    @EqualsAndHashCode.Exclude
     @ManyToOne(optional = false)
     @JoinColumn(name = "person_id", nullable = false)
     private Person person;
 
     @Builder.Default
+    @EqualsAndHashCode.Exclude
     @OrderColumn(name = "ind")
     @ElementCollection(fetch = FetchType.EAGER)
     @CollectionTable(name = "reg_known_person", joinColumns = @JoinColumn(name = "registration_id"))
