+++ src/main/java/sk/leziemevpezinku/spring/model/enums/EnumerationValues.java
@@ -1,8 +1,8 @@
 package sk.leziemevpezinku.spring.model.enums;
 
-public class EnumerationValues {
+public interface EnumerationValues {
 
-    public static class REG_E_EVENT_DISCOUNT_TYPE {
-        public static final String LETO_TABOR_25 = "25TABOR";
+    interface REG_E_EVENT_DISCOUNT_TYPE {
+        String LETO_TABOR_25 = "25TABOR";
     }
 }
