+++ src/main/java/sk/leziemevpezinku/spring/repo/EventTermRepository.java
@@ -2,6 +2,7 @@ package sk.leziemevpezinku.spring.repo;
 
 import jakarta.persistence.LockModeType;
 import org.springframework.data.jpa.repository.EntityGraph;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Lock;
 import org.springframework.data.repository.ListCrudRepository;
@@ -11,7 +12,7 @@ import sk.leziemevpezinku.spring.model.EventTerm;
 import java.util.Optional;
 
 @Repository
-public interface EventTermRepository extends ListCrudRepository<EventTerm, Long>, JpaSpecificationExecutor<EventTerm> {
+public interface EventTermRepository extends JpaRepository<EventTerm, Long>, JpaSpecificationExecutor<EventTerm> {
 
     @EntityGraph(type = EntityGraph.EntityGraphType.LOAD, attributePaths = {"registrations"})
     Optional<EventTerm> findLoadedById(Long id);
