+++ src/main/java/sk/leziemevpezinku/spring/repo/PlaceRepository.java
@@ -1,0 +1,9 @@
+package sk.leziemevpezinku.spring.repo;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import sk.leziemevpezinku.spring.model.Place;
+
+@Repository
+public interface PlaceRepository extends JpaRepository<Place, Long> {
+}
