+++ src/main/java/sk/leziemevpezinku/spring/service/impl/EnumerationServiceBean.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
-import sk.leziemevpezinku.spring.model.enums.EnumerationItem;
-import sk.leziemevpezinku.spring.model.enums.EnumerationItem_;
+import sk.leziemevpezinku.spring.model.EnumerationItem;
+import sk.leziemevpezinku.spring.model.EnumerationItem_;
 import sk.leziemevpezinku.spring.model.enums.EnumerationName;
 import sk.leziemevpezinku.spring.repo.EnumerationRepository;
 import sk.leziemevpezinku.spring.service.EnumerationService;
